Changes AQI value to use the EPA conversion formula
Existing sensor shows the Purple AIr's AQI calculation, this update changes the sensor to return an AQI value using the EPA's conversion formula. The EPA AQI conversion results in a value which is more in line with what local government and media outlets reference. The EPA conversion also better accounts for woodsmoke.

This is a rough first pass at just getting the EPA conversion data in to Home Assistant.